twitter/requirements.txt
absl-py==0.1.10
arrow==0.12.1
asn1crypto==0.24.0
aws-xray-sdk==0.95
awscli==1.14.32
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.5.22
botocore==1.8.36
bs4==0.0.1
bson==0.5.2
bz2file==0.98
cachetools==2.0.1
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
colorama==0.3.7
cookies==2.2.1
credentials==1.1
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
Cython==0.28.1
decorator==4.3.0
dill==0.2.7.1
docker==3.0.0
docker-pycreds==0.2.1
docutils==0.14
feedparser==5.2.1
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gensim==3.3.0
google==2.0.1
google-api-core==0.1.4
google-auth==1.3.0
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.0
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.9.0
h5py==2.7.1
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
ipython-genutils==0.2.0
Jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.1
jsonpickle==0.9.5
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.1.6
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mock==2.0.0
moto==1.2.0
nbformat==4.4.0
nltk==3.2.5
numpy==1.14.0
oauth2client==3.0.0
oauthlib==2.0.6
pandas==0.22.0
pbr==3.1.1
plotly==2.6.0
ply==3.8
prettytable==0.7.2
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.1
psutil==5.4.3
pyaml==17.12.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyparsing==2.2.0
pyquery==1.4.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
smart-open==1.5.6
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
textblob==0.15.1
TFANN==1.0.1
traitlets==4.3.2
tweepy==3.5.0
twitter==1.18.0
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.46.0
Werkzeug==0.14.1
wrapt==1.10.11
xmltodict==0.11.0
